+++ common/src/main/java/com/st0x0ef/stellaris/common/menus/WaterSeparatorMenu.java
@@ -23,7 +23,7 @@ public class WaterSeparatorMenu extends BaseContainer {
     }
 
     public WaterSeparatorMenu(int containerId, Inventory inventory, Container container, WaterSeparatorBlockEntity blockEntity) {
-        super(MenuTypesRegistry.WATER_SEPARATOR_MENU.get(), containerId, 4, inventory, 8, 142);
+        super(MenuTypesRegistry.WATER_SEPARATOR_MENU.get(), containerId, 4, inventory, 10, 142);
         this.container = container;
         this.blockEntity = blockEntity;